requirements.txt
streamlit
pinecone-client
langchain
langchain-huggingface
sentence-transformers
groq
langchain-groq
langchain-community
